Move src/pharmpy/results.py to src/pharmpy/workflows/results.py
